refactor(etl): move method to proper module to enforce package responsibility
